requirements-dev.txt
pre-commit
black
autoflake
pyupgrade
reorder-python-imports
yesqa
pytest-django
pytest-factoryboy